refactor: bigint user id

+++ db/migrate/20260121170436_create_rodauth_email_auth.rb
@@ -1,7 +1,7 @@
 class CreateRodauthEmailAuth < ActiveRecord::Migration[8.1]
   def change
     create_table :user_email_auth_keys, id: false do |t|
-      t.integer :id, primary_key: true
+      t.bigint :id, primary_key: true
       t.foreign_key :users, column: :id
       t.string :key, null: false
       t.datetime :deadline, null: false
